# packages/shared-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Site des Plan\u00e8tes API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/google": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Google Login",
        "description": "Redirige vers Google OAuth. Origin encod\u00e9 et sign\u00e9 (HMAC) dans le state CSRF.",
        "operationId": "google_login_api_v1_auth_google_get",
        "parameters": [
          {
            "name": "origin",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "'frontend' ou 'backoffice'",
              "default": "frontend",
              "title": "Origin"
            },
            "description": "'frontend' ou 'backoffice'"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/google/callback": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Google Callback",
        "description": "Callback Google OAuth.\n- V\u00e9rifie la signature CSRF du state\n- \u00c9change le code contre les infos utilisateur\n- V\u00e9rifie le domaine @eleven-labs.com\n- Cr\u00e9e l'astronaute si premi\u00e8re connexion\n- Redirige vers frontend ou backoffice avec le JWT\nNote: le JWT est transmis en query param (?token=) \u2014 trade-off accept\u00e9 pour ce flow\nde dev (CLAUDE.md \u00a75). Migration vers httpOnly cookie pr\u00e9vue.",
        "operationId": "google_callback_api_v1_auth_google_callback_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Code OAuth retourn\u00e9 par Google",
              "title": "Code"
            },
            "description": "Code OAuth retourn\u00e9 par Google"
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "State CSRF sign\u00e9 (HMAC)",
              "title": "State"
            },
            "description": "State CSRF sign\u00e9 (HMAC)"
          },
          {
            "name": "accept",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "text/html",
              "title": "Accept"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Get Me",
        "description": "Retourne le profil de l'astronaute connect\u00e9.",
        "operationId": "get_me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AstronautMe"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/astronauts": {
      "get": {
        "tags": [
          "astronauts"
        ],
        "summary": "List Astronauts",
        "operationId": "list_astronauts_api_v1_astronauts_get",
        "parameters": [
          {
            "name": "planet_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrer par plan\u00e8te",
              "title": "Planet Id"
            },
            "description": "Filtrer par plan\u00e8te"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AstronautOut"
                  },
                  "title": "Response List Astronauts Api V1 Astronauts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/astronauts/{astronaut_id}": {
      "get": {
        "tags": [
          "astronauts"
        ],
        "summary": "Get Astronaut",
        "operationId": "get_astronaut_api_v1_astronauts__astronaut_id__get",
        "parameters": [
          {
            "name": "astronaut_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Astronaut Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AstronautOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "astronauts"
        ],
        "summary": "Update Astronaut Profile",
        "description": "Met \u00e0 jour le profil d'un astronaute.\nL'astronaute ne peut modifier que son propre profil (photo_url, hobbies, client).\nUn admin peut en plus modifier : planet_id, hire_date, first_name, last_name.",
        "operationId": "update_astronaut_profile_api_v1_astronauts__astronaut_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "astronaut_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Astronaut Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AstronautPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AstronautOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/astronauts/{astronaut_id}/roles": {
      "patch": {
        "tags": [
          "astronauts"
        ],
        "summary": "Update Astronaut Roles",
        "description": "Met \u00e0 jour les r\u00f4les d'un astronaute (admin uniquement).\nUn admin ne peut pas se retirer lui-m\u00eame le r\u00f4le admin.",
        "operationId": "update_astronaut_roles_api_v1_astronauts__astronaut_id__roles_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "astronaut_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Astronaut Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AstronautRoleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AstronautOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AstronautMe": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles"
          },
          "planet_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planet Id"
          },
          "total_points": {
            "type": "integer",
            "title": "Total Points"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "first_name",
          "last_name",
          "photo_url",
          "roles",
          "planet_id",
          "total_points"
        ],
        "title": "AstronautMe"
      },
      "AstronautOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "hobbies": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hobbies"
          },
          "client": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client"
          },
          "hire_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hire Date"
          },
          "planet_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planet Id"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles"
          },
          "total_points": {
            "type": "integer",
            "title": "Total Points"
          },
          "grade_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "first_name",
          "last_name",
          "photo_url",
          "hobbies",
          "client",
          "hire_date",
          "planet_id",
          "roles",
          "total_points",
          "created_at",
          "updated_at"
        ],
        "title": "AstronautOut"
      },
      "AstronautPatch": {
        "properties": {
          "photo_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "hobbies": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Hobbies"
          },
          "client": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Client"
          },
          "planet_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planet Id"
          },
          "hire_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hire Date"
          },
          "first_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          }
        },
        "type": "object",
        "title": "AstronautPatch",
        "description": "Body unifi\u00e9 pour PATCH /astronauts/{id}.\nChamps self-edit : photo_url, hobbies, client.\nChamps admin uniquement : planet_id, hire_date, first_name, last_name."
      },
      "AstronautRoleUpdate": {
        "properties": {
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Roles"
          }
        },
        "type": "object",
        "required": [
          "roles"
        ],
        "title": "AstronautRoleUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}